requirements.txt
astroid==2.9.3
async-generator==1.10
attrs==21.4.0
black==22.1.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-WTF==1.0.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
lxml==4.8.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
outcome==1.1.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.0
psycopg2-binary==2.9.3
pycparser==2.21
pylint==2.12.2
pymongo==4.0.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
SQLAlchemy-Paginator==0.2
toml==0.10.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
unicode-slugify==0.1.5
Unidecode==1.3.3
urllib3==1.26.8
Werkzeug==2.0.3
wget==3.2
wrapt==1.13.3
wsproto==1.1.0
WTForms==3.0.1
